VS2010 Console Application
Code to provide the backend on the local computer that is connected to the rotator. It monitors two text files to receive instructions or send information to generate feedback data. The user input is recorded in a text file which is then read by the backend. The backend then translates the input to a string command that tells the rotator to move. The user request to get the antenna's current position is sent to the rotator and then the text file is appended to reflect its current position which is then displayed on the web interface.